server update-inventory: make dryrun not check if already scheduled
Normally when update-inventory is run a check is performed to determine
if an inventory update is already scheduled for the system. If an update
is, the server will respond with a 409 error saying an update is already
scheduled.

However, if dryrun is true the check is still performed. This seems
unnecessary and prevents someone from running the command just to
generate the xml. This change makes it so the check is not performed
when dryrun is True.

Bug: 1335394
Change-Id: I71cc8af01f4087b7b7a250d74d1004a876f4148f
Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>