server: add device and system level firmware reporting
The system level firmware reporting can be obtained either from the user
by manually editing the system firmware fields from the system edit page
or the information can be obtained from running the inventory task on a
machine.

The device level firmware is not changable by the user and is only able
to be modified through the submission of an inventory report to the
server. The way device firmware information is obtained is hidden with
lshw.

Testing Notes:
The only thing that didn't work was support for older beaker-system-scan
versions. If the inventory report didn't contain BIOS firmware
information the server would error out and abort the report submission,
which is a bug.

Change-Id: Ia9592587afd5a0f52360f2bc5a7f017ac068a843
Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>
Signed-Off-By: Shawn Doherty <sdoherty@redhat.com>