efi boot_order: fix order so one doesn't have to use next boot option
This patch is predicated on the assumption that a beaker provisioned
system was netbooted so the first entry in the boot order before
installation of the OS was some kind of netboot entry. Additionally, it
further assumes that the OS installer properly fixed up the boot order
to cause the OS to be the first thing to boot.

Therefore, if we adjust this boot_order script to place the previously
used boot option before install as the first option, and then place the
OS boot option after we have effectivally preserved the waterfall effect
of a non-efi boot order, that is; net -> disk

This also removes the need for some custom rhts-reboot script that no
common user knows about and just wants the OS's "reboot" command to just
work.

Change-Id: I977b0c906391cfb67e4753d7e6c9a85cf7d0cccc
Signed-off-by: Jonathan Toppins <jtoppins@redhat.com>